requirements.txt
asgiref==3.4.1
bleach==4.1.0
certifi==2021.5.30
charset-normalizer==2.0.4
cloudinary==1.29.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.6
Django==3.2.6
django-cloudinary-storage==0.3.0
django-cors-headers==3.8.0
django-filter==21.1
dj-database-url==0.5.0
django-jazzmin==2.6.0
django-js-asset==1.2.2
django-pandas==0.6.6
django-rest-passwordreset==1.2.0
django-summernote==0.8.20.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-yasg==1.20.0
filelock==3.8.0
gunicorn==20.1.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==21.0
Pillow==9.4.0
platformdirs==2.5.2
psycopg2-binary==2.9.1
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
virtualenv==20.16.5
webencodings==0.5.1
whitenoise==5.1.0
